% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.summary.by.period.R
\name{activity.summary.by.custom.period}
\alias{activity.summary.by.custom.period}
\title{Summarise activity data from a folder of activPAL events files
in user defined periods}
\usage{
activity.summary.by.custom.period(
  location_file,
  output_folder,
  minimum_valid_wear = 20
)
}
\arguments{
\item{location_file}{The filepath to a csv file containing the path for a group
of csv files to be processed}

\item{output_folder}{The filepath for the folder where the csv summary files are saved}

\item{minimum_valid_wear}{The minimum duration of valid wear time required for
a day to be included in the analysis.
Valid daily wear time is the total daily wear time, excluding periods of wear
classified as non-wear (activity code - 4)
Default - 20 hours}
}
\description{
activity.summary.custom.period reads in a csv file containing a list
    of events files and csv files with time periods of interest (the first two
    columns should have the start date and end date in the format dd-mm-yy HH:MM).
    A summary table is saved to a csv file for each events file.  A further csv
    file giving a summary for all the events file is also generated.  The generated
    summary activity table for each events file is saved to a csv file, while an
    additional csv file containing the summaries across all the processed events
    files is also saved.
}
