% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.summary.by.period.R
\name{activity.summary.by.custom.period.revised}
\alias{activity.summary.by.custom.period.revised}
\title{Summarise activity data from a folder of activPAL events files
in user defined periods.  This is an updated version of
activity.summary.by.custom.period to support the specification of
custom reporting periods used in generate.physical.behaviour.summary()
and the include details on the number of events occurring within
each epoch.}
\usage{
activity.summary.by.custom.period.revised(
  input_folder,
  period_file = NULL,
  output_folder,
  prefix_delimiter = NULL,
  prefix_length = NULL,
  fill_gaps = FALSE,
  minimum_valid_wear = 20
)
}
\arguments{
\item{input_folder}{The folder where the events files (format *EventsEx.csv)
and optional rise time files (format *RiseSitData.csv) to be processed are saved.}

\item{period_file}{The location of a csv files containing the details of custom periods to
    to be used when analysing the supplied events files.

    If the location of a valid file is supplied the csv file with the suffix "_Custom_Summary"
    will be generated containing the outcomes using the supplied periods.

    Each row defines a single observation for a single events file. The following columns
    must be in the spreadsheet.
    filename = an identifier for the events file that the period is for.  This should be
    the same as the identifier generated using the prefix_delimiter and prefix_length parameters
    startime = the starting time of the period
    endtime = the ending time of the period
    The function will attempt to parse the supplied times in the following order
    "YYYY-mm-dd HH:MM", "YYYY-mm-dd HH:MM:SS", "YYYY-dd-mm HH:MM", "YYYY-dd-mm HH:MM:SS",
    "dd/mm/YYYY HH:MM", "dd/mm/YYYY HH:MM:SS", "mm/dd/YYYY HH:MM", "mm/dd/YYYY HH:MM:SS"}

\item{output_folder}{The folder where the physical behaviour summary report is to be saved.
By default the report is saved in the current working directory}

\item{prefix_delimiter}{Character string to be matched against the events file name when
generating the file identifier.  The generated identifier is the portion of the file name
preceding the supplied string.}

\item{prefix_length}{An integer that specifies the number of characters to take from the
start of the events file name to generate the file identifier.
If prefix_delimiter and prefix_length are both provided, the shortest file identifier
that can be generated using the parameters is used.
If neither prefix_delimiter or prefix_length are provided, the first six characters
of the file name is used as the file identifier.}

\item{fill_gaps}{Specifies if outcomes should be calculated for any time periods not
explicitly specified by a period file.
This parameter will only be used if a valid period file is supplied}

\item{minimum_wear_time}{the minimum number of hours of valid wear}
}
\description{
activity.summary.custom.period.revised reads in a csv file containing a list
    of events files and csv files with time periods of interest (the first two
    columns should have the start date and end date in the format dd-mm-yy HH:MM).
    A summary table is saved to a csv file for each events file.  A further csv
    file giving a summary for all the events file is also generated.  The generated
    summary activity table for each events file is saved to a csv file, while an
    additional csv file containing the summaries across all the processed events
    files is also saved.
}
