% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Rcpp.R
\name{activpal.calculate.cadence.change}
\alias{activpal.calculate.cadence.change}
\title{Title activpal.calculate.cadence.change}
\usage{
activpal.calculate.cadence.change(data)
}
\arguments{
\item{data}{}
}
\description{
Title activpal.calculate.cadence.change
}
