% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.events.file.R
\name{activpal.convert.events.extended.file}
\alias{activpal.convert.events.extended.file}
\title{Title activpal.convert.events.extended.file}
\usage{
activpal.convert.events.extended.file(data)
}
\arguments{
\item{data}{}
}
\description{
Title activpal.convert.events.extended.file
}
