% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.events.file.R
\name{activpal.file.process}
\alias{activpal.file.process}
\title{Title activpal.file.process}
\usage{
activpal.file.process(data, valid.days = NULL, wear.time.minimum = 72000)
}
\arguments{
\item{wear.time.minimum}{}
}
\description{
Title activpal.file.process
}
