% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.events.file.R
\name{activpal.file.process.rename.row}
\alias{activpal.file.process.rename.row}
\title{Title activpal.file.process.rename.row}
\usage{
activpal.file.process.rename.row(data)
}
\arguments{
\item{data}{}
}
\description{
Title activpal.file.process.rename.row
}
