% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{activpal.process.folder.macbook}
\alias{activpal.process.folder.macbook}
\title{Title activpal.process.folder.macbook}
\usage{
activpal.process.folder.macbook(
  Events_Files_To_Process_folder_location,
  Temp_Output_folder_location,
  Confirmed_Output_folder_location
)
}
\arguments{
\item{Temp_Output_folder_location}{}
}
\description{
Title activpal.process.folder.macbook
}
