% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_special_windows.R
\name{activpal.process.folder.windows}
\alias{activpal.process.folder.windows}
\title{Title activpal.process.folder.windows}
\usage{
activpal.process.folder.windows(
  Events_Files_To_Process_folder_location,
  Temp_Output_folder_location,
  Confirmed_Output_folder_location
)
}
\arguments{
\item{Temp_Output_folder_location}{}
}
\value{
current_batched_ids
}
\description{
Title activpal.process.folder.windows
}
