% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Rcpp.R
\name{activpal.remove.longer.bouts}
\alias{activpal.remove.longer.bouts}
\title{Title activpal.remove.longer.bouts}
\usage{
activpal.remove.longer.bouts(file_data, lower_bout_length, upper_bout_length)
}
\arguments{
\item{upper_bout_length}{}
}
\description{
Title activpal.remove.longer.bouts
}
