% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Rcpp.R
\name{activpal.stepping.process.file}
\alias{activpal.stepping.process.file}
\title{Title activpal.stepping.process.file}
\usage{
activpal.stepping.process.file(
  events_file,
  window_size = 2700,
  lower_bout_size = 0.1,
  max_bout_size = 86400,
  wear_time_minimum = 72000,
  daily_summary = TRUE
)
}
\arguments{
\item{daily_summary}{}
}
\description{
Title activpal.stepping.process.file
}
