% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Rcpp.R
\name{activpal.stepping.test.file}
\alias{activpal.stepping.test.file}
\title{Title activpal.stepping.test.file}
\usage{
activpal.stepping.test.file(file.data, window.size)
}
\arguments{
\item{window.size}{}
}
\description{
Title activpal.stepping.test.file
}
