% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.by.period.R
\name{activpal.summary.by.period}
\alias{activpal.summary.by.period}
\title{Generate event file summaries from a folder of activPAL events files}
\usage{
activpal.summary.by.period(
  location_folder,
  output_folder,
  window_duration = 24
)
}
\arguments{
\item{location_folder}{The filepath for the folder where the events files to be processed
are saved}

\item{output_folder}{The filepath for the folder where the csv summary files are saved}

\item{window_duration}{The size of each bucket in hours that the data is to be summarised.
Default - 24 hours.}
}
\description{
activity.summary.by.period processes all the events files (format *Events.csv)
    in a folder and produces a summary table giving a breakdown across consecutive
    periods of a specified duration during the day. The output generated matches the
    output format obtained from using the Daily summaries export in PALanalysis / PALbatch

    The generated summary activity table for each events file is saved to a csv file,
    while an additional csv file containing the summaries across all the processed
    events files is also saved.

    If the window duration is not a factor of 24, some summary period may encompass events
    spanning two distinct days
}
