% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{apSummary.macbook}
\alias{apSummary.macbook}
\title{Title apSummary.macbook}
\usage{
apSummary.macbook(
  Events_Files_To_Process_folder_location,
  Confirmed_Output_folder_location
)
}
\arguments{
\item{Confirmed_Output_folder_location}{}
}
\description{
Title apSummary.macbook
}
