% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{fun_call}
\alias{fun_call}
\title{Title Check for functions used in a function}
\usage{
fun_call(f)
}
\arguments{
\item{f}{the interested function}
}
\value{
functions which are used in the interested function
}
\description{
Title Check for functions used in a function
}
