% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activpal.physical.behaviour.summary.R
\name{generate.physical.behaviour.summary}
\alias{generate.physical.behaviour.summary}
\title{Generates a PDF that visualises a range of summary physical behaviour outcomes for a group of individuals}
\usage{
generate.physical.behaviour.summary(
  input_folder,
  period_file = NULL,
  chart_title,
  output_folder = "",
  prefix_delimiter = NULL,
  prefix_length = NULL,
  fill_gaps = FALSE,
  minimum_wear_time = 20,
  anonymise = FALSE,
  standard_scales = FALSE,
  sort_order = "MEDIAN_DAILY_STEP_COUNT"
)
}
\arguments{
\item{input_folder}{The folder where the events files (format *EventsEx.csv)
and optional rise time files (format *RiseSitData.csv) to be processed are saved.}

\item{period_file}{The location of a csv files containing the details of custom periods to
    to be used when analysing the supplied events files.

    If the location of a valid file is supplied the csv file with the suffix "_Custom_Summary"
    will be generated containing the outcomes using the supplied periods.

    Each row defines a single observation for a single events file. The following columns
    must be in the spreadsheet.
    filename = an identifier for the events file that the period is for.  This should be
    the same as the identifier generated using the prefix_delimiter and prefix_length parameters
    startime = the starting time of the period
    endtime = the ending time of the period
    The function will attempt to parse the supplied times in the following order
    "YYYY-mm-dd HH:MM", "YYYY-mm-dd HH:MM:SS", "YYYY-dd-mm HH:MM", "YYYY-dd-mm HH:MM:SS",
    "dd/mm/YYYY HH:MM", "dd/mm/YYYY HH:MM:SS", "mm/dd/YYYY HH:MM", "mm/dd/YYYY HH:MM:SS"}

\item{chart_title}{Character string; the file identifier for the generated PDF file.}

\item{output_folder}{The folder where the physical behaviour summary report is to be saved.
By default the report is saved in the current working directory}

\item{prefix_delimiter}{Character string to be matched against the events file name when
generating the file identifier.  The generated identifier is the portion of the file name
preceding the supplied string.}

\item{prefix_length}{An integer that specifies the number of characters to take from the
start of the events file name to generate the file identifier.
If prefix_delimiter and prefix_length are both provided, the shortest file identifier
that can be generated using the parameters is used.
If neither prefix_delimiter or prefix_length are provided, the first six characters
of the file name is used as the file identifier.}

\item{fill_gaps}{Specifies if outcomes should be calculated for any time periods not
explicitly specified by a period file.
This parameter will only be used if a valid period file is supplied}

\item{minimum_wear_time}{the minimum number of hours of valid wear}

\item{anonymise}{logical; if true, set the file identifier as a non-identifiable string.}

\item{standard_scales}{logical; if true, time-based axis limits are set at 24 hours.
Using standard scales is likely to result in significant amounts of empty space.}

\item{sort_order}{character string; selects the outcome measure that is used to
sort the observations.
By default, files are sorted by decreasing mean daily sedentary time.}
}
\description{
activity.summary.window processes all the extended events files (format *EventsEx.csv)
    in a folder and produces a chart showing a number of summary physical behaviour outcomes for
    the files.  The outcomes are grouped into three groups:
    1. Mean daily time in different posture classes (time in bed, sedentary, upright and stepping)
    2. Volume based measures of activity participation (step count, stepping duration and intensity,
       time spent in different travel associated activities)
    3. Measures of physical ability (time to first step, maximum step count, median cadence of
       stepping bouts containing short (< 1 minute) and long (> 1 minute) periods of stepping)

    If matching rise time files (format *RiseSitData.csv) are included in the folder an additional
    column will be inserted to show the median rise time.
}
\details{
The following outcomes can be used to sort observations in the chart
    SEDENTARY_TIME, TIME_IN_BED, UPRIGHT_TIME, STEPPING_TIME,
    DAILY_SEDENTARY_BOUTS, DAILY_TIB_INTERRUPTIONS, STEPPING_BOUTS_UNDER_1_MIN,
    PEAK_2_MIN_STEPPING, PEAK_6_MIN_STEPPING, PEAK_12_MIN_STEPPING,
    DAILY_PEAK_30_SECOND_STEPPING, MEDIAN_DAILY_STEP_COUNT, MEDIAN_RISE_TIME,
    MEDIAN_CADENCE_TO_1_MINUTE , MEDIAN_CADENCE_1_MINUTE_PLUS, STEPPING_INTENSITY
}
