% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{make.index.file.macbook}
\alias{make.index.file.macbook}
\title{Title make.index.file.macbook}
\usage{
make.index.file.macbook(
  Events_Files_To_Process_folder_location,
  Temp_Output_folder_location
)
}
\arguments{
\item{Temp_Output_folder_location}{}
}
\description{
Title make.index.file.macbook
}
