% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.events.file.R
\name{pre.process.events.file}
\alias{pre.process.events.file}
\title{Title pre.process.events.file}
\usage{
pre.process.events.file(file_name, folder = "", minimum_valid_wear = 20)
}
\arguments{
\item{minimum_valid_wear}{}
}
\description{
Title pre.process.events.file
}
