% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{prepare.ex.times.macbook}
\alias{prepare.ex.times.macbook}
\title{Title prepare.ex.times.macbook}
\usage{
prepare.ex.times.macbook(Temp_Output_folder_location)
}
\arguments{
\item{Temp_Output_folder_location}{}
}
\description{
Title prepare.ex.times.macbook
}
