% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{prepare.ex.times.windows}
\alias{prepare.ex.times.windows}
\title{Title prepare.ex.times.windows}
\usage{
prepare.ex.times.windows(Temp_Output_folder_location)
}
\arguments{
\item{Temp_Output_folder_location}{}
}
\description{
Title prepare.ex.times.windows
}
