% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{sed.activity.marker}
\alias{sed.activity.marker}
\title{Title sed.activity.marker}
\usage{
sed.activity.marker(events_file, time_value, duration)
}
\arguments{
\item{duration}{}
}
\description{
Title sed.activity.marker
}
