% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{sed.start.end.marker}
\alias{sed.start.end.marker}
\title{Title sed.start.end.marker}
\usage{
sed.start.end.marker(
  events_file,
  time_value,
  event = 0,
  duration.min = 30,
  duration.max = 600,
  interval = 600
)
}
\arguments{
\item{interval}{}
}
\description{
Title sed.start.end.marker
}
