% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{stand.start.end.marker}
\alias{stand.start.end.marker}
\title{Title stand.start.end.marker}
\usage{
stand.start.end.marker(
  events_file,
  time_value,
  event = 1,
  duration.min = 30,
  duration.max = 600,
  interval = 600
)
}
\arguments{
\item{interval}{}
}
\description{
Title stand.start.end.marker
}
