% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{transportation.activity.marker}
\alias{transportation.activity.marker}
\title{Title transportation.activity.marker}
\usage{
transportation.activity.marker(events_file, time_value, duration)
}
\arguments{
\item{duration}{}
}
\description{
Title transportation.activity.marker
}
