% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{activpal.process.folder.windows}
\alias{activpal.process.folder.windows}
\title{Title 02 Identify Possible Exercise Bouts}
\usage{
activpal.process.folder.windows(
  Events_Files_To_Process_folder_location,
  Temp_Output_folder_location
)
}
\arguments{
\item{Temp_Output_folder_location}{}
}
\value{
current_batched_ids
}
\description{
Title 02 Identify Possible Exercise Bouts
}
